Move GC name to be specific to Functions.
