+++ engine/src/conversion/utilities.rs
@@ -35,7 +35,6 @@ pub(crate) fn generate_utilities(apis: &mut Vec<UnanalyzedApi>) {
         use_stmt: Use::Unused,
         deps: HashSet::new(),
         detail: super::api::ApiDetail::StringConstructor,
-        id_for_allowlist: None,
         additional_cpp: Some(AdditionalNeed::MakeStringConstructor),
     });
 }
